Move new_generics_with_where_clause to new file utils.rs
