+++ doku-derive/src/expand.rs
@@ -5,12 +5,13 @@ mod expand_struct;
 mod expand_variant;
 mod expand_variants;
 mod expand_wrap;
+mod utils;
 
 use self::{
     expand_enum::expand_enum, expand_field::expand_field,
     expand_fields::expand_fields, expand_struct::expand_struct,
     expand_variant::expand_variant, expand_variants::expand_variants,
-    expand_wrap::expand_wrap,
+    expand_wrap::expand_wrap, utils::*,
 };
 use crate::prelude::*;
 
@@ -24,23 +25,4 @@ pub fn expand(input: &syn::DeriveInput) -> Result<TokenStream2> {
         )
         .into()),
     }
-}
-
-fn new_generics_with_where_clause(
-    generics: &syn::Generics,
-) -> Result<syn::Generics> {
-    let mut new_generics = generics.to_owned();
-    let where_clause = new_generics.make_where_clause();
-    for generic in &generics.params {
-        match generic {
-            syn::GenericParam::Const(_) => (),
-            syn::GenericParam::Lifetime(_) => (),
-            syn::GenericParam::Type(t) => {
-                let predicate: syn::WherePredicate =
-                    syn::parse2(quote! { #t: ::doku::Document }).unwrap();
-                where_clause.predicates.push(predicate);
-            }
-        };
-    }
-    Ok(new_generics)
 }
