+++ doku-derive/src/expand/utils.rs
@@ -1,0 +1,20 @@
+use crate::prelude::*;
+
+pub fn new_generics_with_where_clause(
+    generics: &syn::Generics,
+) -> Result<syn::Generics> {
+    let mut new_generics = generics.to_owned();
+    let where_clause = new_generics.make_where_clause();
+    for generic in &generics.params {
+        match generic {
+            syn::GenericParam::Const(_) => (),
+            syn::GenericParam::Lifetime(_) => (),
+            syn::GenericParam::Type(t) => {
+                let predicate: syn::WherePredicate =
+                    syn::parse2(quote! { #t: ::doku::Document }).unwrap();
+                where_clause.predicates.push(predicate);
+            }
+        };
+    }
+    Ok(new_generics)
+}
